HealthCare_backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
pydantic

